+++ routes/api.php
@@ -81,6 +81,7 @@ Route::get('products', 'Api\ProductController@index');
 Route::get('products/{id}', 'Api\ProductController@show');
 Route::get('search-products', 'Api\ProductController@search');
 Route::get('product-with-settings/{id}', 'Api\ProductController@showProductWithSettings');
+Route::get('popular-products', 'Api\ProductController@getPopularProducts');
 
 Route::get('product-categories', 'Api\ProductCategoryController@index');
 Route::get('product-categories/{id}', 'Api\ProductCategoryController@show');